+++ src/checker/expr/mod.rs
@@ -3,5 +3,6 @@
 use super::*;
 
 mod core;
+mod lambda;
 mod literals;
 mod operators;